minor changes to make eslint happy
