--- wp-content/themes/accelerate/podcasting-platform/components/PlayButton/templates/initial.html
@@ -1,7 +1,0 @@
-<button class="ilo-button ilo-button--outlined yourRippleEffectClass">
-    <svg class="ilo-button-icon" width="24" height="24" viewBox="0 0 24 24" aria-label="play podcast">
-      <path id="arc1" fill="none" stroke="#0078D4" stroke-width="1.82" d="M 11.841357625485093 2.911384451028404 A 9.09 9.09 0 1 0 12 2.91"></path>
-      <path d="M15.6359 11.9998L10.1814 15.9362V8.06348L15.6359 11.9998Z" fill="#0078D4"></path>
-    </svg>
-    <span class="ilo-button__label">enabled</span>
-</button>
