--- wp-content/themes/accelerate/podcasting-platform/components/PlayButton/templates/playing.html
@@ -1,10 +1,0 @@
-<button class="ilo-button ilo-button--outlined yourRippleEffectClass">
-    <svg id="svgSound-playing0" class="svgSound-playing" style="display: block;" height="24px" stroke-width="4" viewBox="-12 -12 24 24">
-        <g jsname="HGYFec">
-          <line class="leftLine" x1="-6" x2="-6" y1="8" y2="-8"></line>
-          <line class="middleLine" x1="0" x2="0" y1="8" y2="-8"></line>
-          <line class="rightLine" x1="6" x2="6" y1="8" y2="-8"></line>
-        </g>
-    </svg>
-    <span class="ilo-button__label"> reproduciendo </span>
-</button>
